# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LunarMood API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthRequest {
  username: string;
  password: string;
}

export interface AuthResponse {
  id: number;
  username: string;
  token: string;
}

export interface User {
  id: number;
  username: string;
}

export interface UserProfile {
  id: number;
  username: string;
  /** @nullable */
  consumptionLabel?: string | null;
}

export interface UpdateProfileRequest {
  consumptionLabel?: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type Period = (typeof Period)[keyof typeof Period];

export const Period = {
  morning: "morning",
  afternoon: "afternoon",
  evening: "evening",
} as const;

export interface MoodEntry {
  id: number;
  userId: number;
  date: string;
  period: Period;
  /**
   * @minimum 1
   * @maximum 5
   */
  mood: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  energy: number;
  /**
   * @minimum 0
   * @maximum 5
   */
  consumption: number;
  /** @nullable */
  note?: string | null;
  lunarPhase: string;
  createdAt: string;
}

export interface CreateMoodRequest {
  date: string;
  period?: Period;
  /**
   * @minimum 1
   * @maximum 5
   */
  mood: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  energy?: number;
  /**
   * @minimum 0
   * @maximum 5
   */
  consumption?: number;
  /** @nullable */
  note?: string | null;
}

export interface UpdateMoodRequest {
  /**
   * @minimum 1
   * @maximum 5
   */
  mood?: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  energy?: number;
  /**
   * @minimum 0
   * @maximum 5
   */
  consumption?: number;
  /** @nullable */
  note?: string | null;
}

export type LunarPhasePhase =
  (typeof LunarPhasePhase)[keyof typeof LunarPhasePhase];

export const LunarPhasePhase = {
  new_moon: "new_moon",
  waxing_crescent: "waxing_crescent",
  first_quarter: "first_quarter",
  waxing_gibbous: "waxing_gibbous",
  full_moon: "full_moon",
  waning_gibbous: "waning_gibbous",
  last_quarter: "last_quarter",
  waning_crescent: "waning_crescent",
} as const;

export interface LunarPhase {
  date: string;
  phase: LunarPhasePhase;
  illumination: number;
  emoji: string;
}

export type ListMoodsParams = {
  /**
   * Month (1-12)
   */
  month?: number;
  /**
   * Year
   */
  year?: number;
};

export type GetLunarPhasesParams = {
  month: number;
  year: number;
};