# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LunarMood API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: moods
    description: Mood entry operations
  - name: lunar
    description: Lunar phase operations
  - name: user
    description: User profile operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Log out
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /user/profile:
    patch:
      operationId: updateProfile
      tags: [user]
      summary: Update user profile settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /moods:
    get:
      operationId: listMoods
      tags: [moods]
      summary: List mood entries for current user
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: integer
          description: Month (1-12)
        - name: year
          in: query
          required: false
          schema:
            type: integer
          description: Year
      responses:
        "200":
          description: List of mood entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MoodEntry"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createMood
      tags: [moods]
      summary: Create a mood entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMoodRequest"
      responses:
        "201":
          description: Mood entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoodEntry"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /moods/{id}:
    patch:
      operationId: updateMood
      tags: [moods]
      summary: Update a mood entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMoodRequest"
      responses:
        "200":
          description: Updated mood entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoodEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMood
      tags: [moods]
      summary: Delete a mood entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /lunar/phases:
    get:
      operationId: getLunarPhases
      tags: [lunar]
      summary: Get lunar phases for a month
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: integer
        - name: year
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lunar phases for the month
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LunarPhase"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AuthRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AuthResponse:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        token:
          type: string
      required:
        - id
        - username
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
      required:
        - id
        - username
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        consumptionLabel:
          type: ["string", "null"]
      required:
        - id
        - username
    UpdateProfileRequest:
      type: object
      properties:
        consumptionLabel:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    Period:
      type: string
      enum:
        - morning
        - afternoon
        - evening
    MoodEntry:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        date:
          type: string
        period:
          $ref: "#/components/schemas/Period"
        mood:
          type: integer
          minimum: 1
          maximum: 5
        energy:
          type: integer
          minimum: 0
          maximum: 100
        consumption:
          type: integer
          minimum: 0
          maximum: 5
        note:
          type: ["string", "null"]
        lunarPhase:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - date
        - period
        - mood
        - energy
        - consumption
        - lunarPhase
        - createdAt
    CreateMoodRequest:
      type: object
      properties:
        date:
          type: string
        period:
          $ref: "#/components/schemas/Period"
        mood:
          type: integer
          minimum: 1
          maximum: 5
        energy:
          type: integer
          minimum: 0
          maximum: 100
        consumption:
          type: integer
          minimum: 0
          maximum: 5
        note:
          type: ["string", "null"]
      required:
        - date
        - mood
    UpdateMoodRequest:
      type: object
      properties:
        mood:
          type: integer
          minimum: 1
          maximum: 5
        energy:
          type: integer
          minimum: 0
          maximum: 100
        consumption:
          type: integer
          minimum: 0
          maximum: 5
        note:
          type: ["string", "null"]
    LunarPhase:
      type: object
      properties:
        date:
          type: string
        phase:
          type: string
          enum:
            - new_moon
            - waxing_crescent
            - first_quarter
            - waxing_gibbous
            - full_moon
            - waning_gibbous
            - last_quarter
            - waning_crescent
        illumination:
          type: number
        emoji:
          type: string
      required:
        - date
        - phase
        - illumination
        - emoji