# lib/api-zod/src/generated/types/createMoodRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LunarMood API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Period } from "./period";

export interface CreateMoodRequest {
  date: string;
  period?: Period;
  /**
   * @minimum 1
   * @maximum 5
   */
  mood: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  energy?: number;
  /**
   * @minimum 0
   * @maximum 5
   */
  consumption?: number;
  /** @nullable */
  note?: string | null;
}