# lib/api-zod/src/generated/types/lunarPhase.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LunarMood API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LunarPhasePhase } from "./lunarPhasePhase";

export interface LunarPhase {
  date: string;
  phase: LunarPhasePhase;
  illumination: number;
  emoji: string;
}