# lib/api-zod/src/generated/types/period.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LunarMood API specification
 * OpenAPI spec version: 0.1.0
 */

export type Period = (typeof Period)[keyof typeof Period];

export const Period = {
  morning: "morning",
  afternoon: "afternoon",
  evening: "evening",
} as const;